requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.14.0
django-allauth>=0.57.0
django-allauth-ui>=0.1.0
django-tailwind>=3.8.0
django-unfold>=0.24.0
django-browser-reload>=1.12.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
python-decouple>=3.8
django-cors-headers>=4.3.1
django-filter>=23.5
django-widget-tweaks>=1.5.0
whitenoise>=6.6.0
gunicorn>=21.2.0

